requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
funcsigs==1.0.2
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.3
pytest==3.6.2
six==1.11.0
selenium
Appium-Python-Client